update aws util
